% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calc_policy_counts}
\alias{calc_policy_counts}
\title{Calculate Policy Counts}
\usage{
calc_policy_counts(file_list, group_name, t_red = 0)
}
\arguments{
\item{file_list}{A list or vector of file paths.}

\item{group_name}{String. The group name.}

\item{t_red}{Integer. Start year for policy filtering.}

\item{var_name}{String. Variable name.}
}
\value{
A dataframe with Group, Mean, and SD.
}
\description{
Calculates statistics for policy periods in simulation results.
}
\keyword{internal}
