% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_design_inputs}
\alias{check_design_inputs}
\title{Check Experiment Design Inputs}
\usage{
check_design_inputs(
  file_obj,
  df,
  esd_str,
  pae_str,
  rm_str,
  breaks_str,
  probs_str,
  comp_mode
)
}
\description{
Validates experiment design CSV and vectors.
}
\keyword{internal}
