% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_num}
\alias{check_num}
\title{Check Numeric Value}
\usage{
check_num(val, name, min_val = -Inf, max_val = Inf, is_int = FALSE)
}
\description{
Validates a numeric input against bounds.
}
\keyword{internal}
