% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_vbgf_inputs}
\alias{check_vbgf_inputs}
\title{Check VBGF Inputs}
\usage{
check_vbgf_inputs(file_obj, df, boot_b)
}
\description{
Validates growth data upload and bootstrap settings.
}
\keyword{internal}
