% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_z_inputs}
\alias{check_z_inputs}
\title{Check Z Estimation Inputs}
\usage{
check_z_inputs(growth_df, full, last, bg2)
}
\description{
Validates mortality estimation parameters.
}
\keyword{internal}
