% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{fill_alk_sd_for_model}
\alias{fill_alk_sd_for_model}
\title{Prepare Automatically Generated ALK Data for the Model}
\usage{
fill_alk_sd_for_model(alk)
}
\arguments{
\item{alk}{A data frame containing Age, n, Length and Lengthsd.}
}
\value{
A model-ready ALK data frame with no missing Lengthsd values.
}
\description{
Replaces undefined within-age standard deviations in an
  automatically generated ALK. The nearest preceding age class with a
  valid SD is preferred. If none exists, the next valid age class is used.
}
\keyword{internal}
