% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_simulation_sizelimit_cpp}
\alias{run_simulation_sizelimit_cpp}
\title{Run IBM Simulation with Size Limits}
\usage{
run_simulation_sizelimit_cpp(
  zr_w_dist,
  month_weights,
  W1_alk,
  agedata,
  harvest_params_in,
  growth_params_dd_in1,
  growth_params_dd_in2,
  survival_params,
  scenario_to_run,
  policy_combos,
  compliance_structure,
  before_policy_years = 50L,
  policy_years = 30L,
  lake_area_ha = 2818.635,
  initial_pop_size = 10000L,
  rec_a = 2.901,
  rec_b = 134.78,
  rec_v = 0.68,
  F_over_Z_ratio = 0.5,
  juv_onlyM_len = 130,
  spawn_month = 4L,
  recruit_entry_month = 8L,
  rep = 1L,
  vmonthly_avg = 0.15,
  min_adult_age = 1,
  age_recruit = 1,
  age_spawn = 1,
  psd_stock = 130,
  psd_quality = 200,
  psd_preferred = 250,
  psd_memorable = 300,
  psd_trophy = 380,
  Fagemode = FALSE,
  use_ricker = TRUE,
  T_safe = 0L
)
}
\description{
Run IBM Simulation with Size Limits
}
