% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{run_z_bootstrap_custom}
\alias{run_z_bootstrap_custom}
\title{Run Z Bootstrap (Custom)}
\usage{
run_z_bootstrap_custom(raw_data, BG2, full, last, method, seed = NULL)
}
\arguments{
\item{raw_data}{Dataframe with 'Age' and 'n' columns.}

\item{BG2}{Integer. Number of bootstrap replicates.}

\item{full}{Integer. Fully recruited age.}

\item{last}{Integer. Maximum age.}

\item{method}{Character. "lr", "wlr", "pois", or "ripois".}
}
\value{
Numeric vector of bootstrapped Z estimates.
}
\description{
Wrapper function to estimate Z using catch curves or GLMs with bootstrapping.
}
\keyword{internal}
