% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{zestimate1_shiny}
\alias{zestimate1_shiny}
\title{Internal Z Estimation Logic}
\usage{
zestimate1_shiny(extd, BG2, method = c("lr", "wlr", "pois", "ripois"))
}
\arguments{
\item{extd}{Data frame with extended age structure.}

\item{BG2}{Number of bootstraps.}

\item{method}{Method string.}
}
\value{
Bootstrapped Z values.
}
\description{
Core logic for Z estimation using car::Boot or lme4.
}
\keyword{internal}
