% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{zestimate_shiny}
\alias{zestimate_shiny}
\title{Internal Z Data Prep}
\usage{
zestimate_shiny(age, number, full, last, BG2, method)
}
\arguments{
\item{age}{Vector of ages.}

\item{number}{Vector of counts.}

\item{full}{Recruitment age.}

\item{last}{Max age.}

\item{BG2}{Bootstraps.}

\item{method}{Method.}
}
\value{
Bootstrapped Z values.
}
\description{
Prepares data for Z estimation.
}
\keyword{internal}
