% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\docType{package}
\name{craibm-package}
\alias{craibm}
\alias{craibm-package}
\title{Fish IBM Helper Functions}
\description{
Helper functions for the craibm package, including VBGF bootstrapping, Z estimation, and batch simulation runners.
}
\author{
\strong{Maintainer}: Developer Name \email{dev@example.com}

}
\keyword{internal}
