% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{parse_and_check_vec}
\alias{parse_and_check_vec}
\title{Parse and Check Numeric Vector}
\usage{
parse_and_check_vec(
  str_val,
  name,
  min_val = -Inf,
  max_val = Inf,
  len_must_be = NULL
)
}
\description{
Parses a string to numeric vector with bounds checking.
}
