% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{run_vbgf_bootstrap_full}
\alias{run_vbgf_bootstrap_full}
\title{Run VBGF Bootstrap}
\usage{
run_vbgf_bootstrap_full(wd, B = 1000, phi_obs = 0.1, seed = NULL)
}
\arguments{
\item{wd}{A dataframe containing the data. Must have columns named 'Length' and 'Age'.}

\item{B}{Integer. The number of bootstrap replicates to run. Default is 1000.}

\item{phi_obs}{Numeric. The observation error parameter. Default is 0.1.}
}
\value{
A list containing 'Theta_clean' (bootstrapped parameters) and 'Data' (original data with fits).
}
\description{
Fits the von Bertalanffy Growth Function (VBGF) using bootstrapping to estimate parameter uncertainty.
}
